# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "A comprehensive REST API for managing a mechanic shop with customers, vehicles, mechanics, service tickets, and inventory. Includes JWT authentication, rate limiting, and caching."
  version: "2.0.0"
  contact:
    name: "API Support"
    url: "https://github.com/KimiCubb/Mechanics-Shop-db"
host: "mechanics-shop-db.onrender.com"
basePath: ""
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "JWT Bearer token. Format: Bearer <token>"
paths:
...
  /inventory:
    post:
      tags:
        - "Inventory"
      summary: "Create a new part"
      description: "Add a new part to inventory"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/PartCreate"
      responses:
        201:
          description: "Part created successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              part:
                $ref: "#/definitions/Part"
        400:
          description: "Invalid input"

    get:
      tags:
        - "Inventory"
      summary: "Get all parts"
      description: "Retrieve all parts in inventory"
      responses:
        200:
          description: "Parts retrieved successfully"
          schema:
            $ref: "#/definitions/PartsListResponse"
  /inventory/{part_id}:
    get:
      tags:
        - "Inventory"
      summary: "Get a specific part"
      description: "Retrieve details for a specific part"
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Part retrieved successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              part:
                $ref: "#/definitions/Part"
        404:
          description: "Part not found"

    put:
      tags:
        - "Inventory"
      summary: "Update a part"
      description: "Update part details such as name and price"
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/PartCreate"
      responses:
        200:
          description: "Part updated successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              part:
                $ref: "#/definitions/Part"
        400:
          description: "Invalid input"
        404:
          description: "Part not found"

    delete:
      tags:
        - "Inventory"
      summary: "Delete a part"
      description: "Remove a part from inventory"
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Part deleted successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
        404:
          description: "Part not found"
  /inventory/search:
    get:
      tags:
        - "Inventory"
      summary: "Search for parts by name"
      description: "Search inventory for parts matching a query string"
      parameters:
        - in: "query"
          name: "q"
          type: "string"
          required: true
          description: "Search query (part name)"
      responses:
        200:
          description: "Search results"
          schema:
            $ref: "#/definitions/PartSearchResponse"
...
definitions:
...
  Part:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      name:
        type: "string"
        example: "Brake Pad"
      price:
        type: "number"
        format: "float"
        example: 45.50
  PartCreate:
    type: "object"
    required:
      - name
      - price
    properties:
      name:
        type: "string"
        minLength: 1
        maxLength: 100
        example: "Brake Pad"
      price:
        type: "number"
        format: "float"
        minimum: 0
        example: 45.50
  PartsListResponse:
    type: "object"
    properties:
      message:
        type: "string"
      count:
        type: "integer"
      parts:
        type: "array"
        items:
          $ref: "#/definitions/Part"
  PartSearchResponse:
    type: "object"
    properties:
      message:
        type: "string"
      query:
        type: "string"
        example: "brake"
      count:
        type: "integer"
      parts:
        type: "array"
        items:
          $ref: "#/definitions/Part"
...